requirements.txt
supervisely==6.23.1
supervisely[apps]==6.23.1
pyyaml==6.0
imgaug==0.4.0
imagecorruptions==1.1.2
opencv-python-headless==3.4.10.35
opencv-python==3.4.10.35
albumentations